Crate spin_telemetry

Re-exports

pub use metrics::HistogramBuckets;

Modules

detector

env

logs

metrics

traces

Macros

counter

Records an increment to the named monotonic counter (as a u64) with the given attributes.

gauge_f64

Records the current value (as an f64) of the named gauge with the given attributes.

gauge_i64

Records the current value (as an i64) of the named gauge with the given attributes.

gauge_u64

Records the current value (as a u64) of the named gauge with the given attributes.

histogram_f64

Records an additional value (as an f64) to the distribution of the named histogram with the given attributes.

histogram_u64

Records an additional value (as a u64) to the distribution of the named histogram with the given attributes.

up_and_down_counter

Records a delta to the named counter (as an i64) with the given attributes.

Functions

extract_trace_context

Extracts the W3C TraceContext from the provided request and sets it as the parent of the current span.

init

Initializes telemetry for Spin using the tracing library.

inject_trace_context

Injects the current W3C TraceContext into the provided request.