spin_telemetry

Module env

Enums

OtlpProtocol

The protocol to use for OTLP exporter.

Functions

otel_logs_enabled

Returns a boolean indicating if the OTEL log layer should be enabled.

otel_metrics_enabled

Returns a boolean indicating if the OTEL metrics layer should be enabled.

otel_sdk_disabled

Returns a boolean indicating if the OTEL SDK should be disabled for all signals.

otel_tracing_enabled

Returns a boolean indicating if the OTEL tracing layer should be enabled.

spin_disable_log_to_tracing

Returns a boolean indicating if the compatibility layer that emits tracing events from applications logs should be disabled.