Crate spin_telemetry

Modules

detector

logs

metrics

traces

Macros

counter

Records an increment to the named counter with the given attributes.

gauge

Records an increment to the named monotonic counter with the given attributes.

histogram

Adds an additional value to the distribution of the named histogram with the given attributes.

monotonic_counter

Records an increment to the named monotonic counter with the given attributes.

Functions

extract_trace_context

Extracts the W3C TraceContext from the provided request and sets it as the parent of the current span.

init

Initializes telemetry for Spin using the tracing library.

inject_trace_context

Injects the current W3C TraceContext into the provided request.